Implement visual timer
