+++ js/game/level-view.js
@@ -1,11 +1,15 @@
 import AbstractView from '../view';
-import {timeMinutes, timeSeconds} from '../util';
+import {timeMinutes, timeSeconds, getVisualTimerCircleLength} from '../timer';
+import {TIME_LIMIT} from '../data/state';
 
 
 export default class LevelView extends AbstractView {
 
   bind() {
-    this.timeElement = this.element.querySelector(`.timer-value`);
+    this.timerElement = this.element.querySelector(`.timer`);
+    this.timerRadius = parseInt(this.timerElement.querySelector(`circle`).getAttribute(`r`), 10);
+
+    this.timeValueElement = this.element.querySelector(`.timer-value`);
   }
 
   renderMistakes(wrongAnswersNumber) {
@@ -17,10 +21,12 @@ export default class LevelView extends AbstractView {
 
   updateTimer(time) {
     if (time < 30) {
-      this.timeElement.classList.add(`timer-value--finished`);
+      this.timeValueElement.classList.add(`timer-value--finished`);
     }
 
-    this.timeElement.querySelector(`.timer-value-mins`).textContent = timeMinutes(time);
-    this.timeElement.querySelector(`.timer-value-secs`).textContent = timeSeconds(time);
+    this.timerElement.setAttribute(`stroke-dashoffset`, getVisualTimerCircleLength(370, TIME_LIMIT, time));
+
+    this.timeValueElement.querySelector(`.timer-value-mins`).textContent = timeMinutes(time);
+    this.timeValueElement.querySelector(`.timer-value-secs`).textContent = timeSeconds(time);
   }
 }
